Use float for horizontal position
This allows for smoother variation in the lateral movement.

type: ini
type: ux

+++ letters.h
@@ -38,7 +38,7 @@ void initialize_dictionary(char *path, reallocator r);
 
 struct word {
 	struct word *next;
-	int x;       /* horizontal coordinate of position */
+	float x;     /* horizontal coordinate of position */
 	int y;       /* vertical coordinate of position */
 	int drop;    /* number of lines to move per turn */
 	int matches; /* Length of matching prefix */
